+++ src/components/10-atoms/input-text/ui.test.js
@@ -227,4 +227,3 @@ test('should fire onInputfieldKeyUp callback on user input', async t => {
 
   await t.expect(await testoutput().innerHTML).eql('x');
 });
-